Listener's Accept() is now threaded.
Added function private to Connection SetBlockingMode(bool blocking); But
it doesn't do anything yet.
Listener is now accepting clients in a thread.
GetNewClient() returns the socket that has connected. -1 if there is no
new client connected. (Can only connect 1 client at a time.)
The simple chat program is not working anymore because of changes in the
server.